+++ bin/create.js
@@ -1,12 +1,12 @@
 #!/usr/bin/env node
 
 import sade from "sade"
-import handler from "./handler.js"
+import { createHandler as handler } from "./handler.js"
 
 sade("create [dir]")
-	.version("v0.3.0")
+	.version("v0.4.0")
 	.describe("quickly create a project template")
 	.example("")
-	.example("test")
+	.example("new")
 	.action(handler)
 	.parse(process.argv)