+++ bin/handler.js
@@ -1,17 +1,16 @@
 import command from "@m4rch/command"
 import { existsSync, readdirSync } from "fs"
-import { join } from "path"
 import { error } from "../src/utils.js"
 
-import options from "./options.js"
-import index from "../src/index.js"
+import { createQuestions } from "./options.js"
+import { create } from "../src/index.js"
 
-export default function handler ( path ) {
-	let dir = join(process.cwd(), path || ".")
+export function createHandler ( path ) {
+	const dir = path || "."
 
 	if (existsSync(dir) && readdirSync(dir).length) return error("can't initiate in a non-empty folder.")
 
-	command(options(dir))
-		.action(index.bind(null, dir))
+	command(createQuestions(dir))
+		.action(create.bind(null, dir))
 		.run()
 }