+++ src/create/misc/extension/test.js
@@ -1,0 +1,13 @@
+export const index = {
+	js: "",
+	ts: "",
+	test: {
+		js: "",
+		ts: ""
+	}
+}
+
+export const runTest = {
+	js: "",
+	ts: ""
+}