+++ src/create/web/index.js
@@ -1,4 +1,3 @@
 import * as impSvelte from "./svelte.js"
 
 export const svelte = ( files, options ) => impSvelte[options.type](files, options)
-