+++ src/create/web/svelte/sveltekit.js
@@ -2,7 +2,6 @@ import * as general from "./sveltekit/general.js"
 import * as misc from "./sveltekit/misc.js"
 import * as templates from "./sveltekit/templates.js"
 import * as typescript from "./sveltekit/typescript.js"
-const withDep = [ "database" ]
 
 export default function sveltekit ( files, options ) {
 	const pkgIndex = files.findIndex(({ name }) => name == "package.json")
@@ -30,10 +29,8 @@ export default function sveltekit ( files, options ) {
 		}
 	}
 
-	if (options.extra.some(( extra ) => withDep.includes(extra)) || options.fonts.length) pkg.dependencies = {}
-
 	options.fonts.forEach(( font ) => {
-		pkg.dependencies[`@fontsource/${font}`] = "^4.5.0"
+		pkg.devDependencies[`@fontsource/${font}`] = "^4.5.0"
 	})
 
 	files[pkgIndex].content = pkg
@@ -48,7 +45,7 @@ export default function sveltekit ( files, options ) {
 		},
 		{
 			name: "src",
-			files: [
+			children: [
 				{
 					name: "app.css",
 					content: general.css(options.fonts)
@@ -59,7 +56,7 @@ export default function sveltekit ( files, options ) {
 				},
 				{
 					name: "routes",
-					files: [
+					children: [
 						{
 							name: "__layout.svelte",
 							content: general.__layout(options.typescript)
@@ -72,13 +69,13 @@ export default function sveltekit ( files, options ) {
 				},
 				{
 					name: "lib",
-					files: []
+					children: []
 				}
 			]
 		},
 		{
 			name: "static",
-			files: [
+			children: [
 				{
 					name: "robots.txt",
 					content: general.robots
@@ -90,7 +87,7 @@ export default function sveltekit ( files, options ) {
 	if (options.typescript) {
 		const srcIndex = files.findIndex(({ name }) => name == "src")
 
-		files[srcIndex].files.push({
+		files[srcIndex].children.push({
 			name: "global.d.ts",
 			content: typescript.global
 		})
@@ -103,20 +100,20 @@ export default function sveltekit ( files, options ) {
 
 	if (options.templates.length) {
 		const srcIndex = files.findIndex(({ name }) => name == "src")
-		const srcLibIndex = files[srcIndex].files.findIndex(({ name }) => name == "lib")
-		const srcRoutesIndex = files[srcIndex].files.findIndex(({ name }) => name == "routes")
+		const srcLibIndex = files[srcIndex].children.findIndex(({ name }) => name == "lib")
+		const srcRoutesIndex = files[srcIndex].children.findIndex(({ name }) => name == "routes")
 
 		options.templates.forEach(( template ) => {
 			switch (template) {
 				case "__error": {
-					files[srcIndex].files[srcRoutesIndex].files.push({
+					files[srcIndex].children[srcRoutesIndex].children.push({
 						name: "__error.svelte",
 						content: templates.__error(options.typescript)
 					})
 					break
 				}
 				case "message": {
-					files[srcIndex].files[srcLibIndex].files.push({
+					files[srcIndex].children[srcLibIndex].children.push({
 						name: "message.svelte",
 						content: templates.message.svelte(options.typescript)
 					},
@@ -131,46 +128,90 @@ export default function sveltekit ( files, options ) {
 
 	if (options.extra.length) {
 		const srcIndex = files.findIndex(({ name }) => name == "src")
-		const srcLibIndex = files[srcIndex].files.findIndex(({ name }) => name == "lib")
-		const pkg = files.findIndex(({ name }) => name == "package.json")
+		const srcLibIndex = files[srcIndex].children.findIndex(({ name }) => name == "lib")
+		const pkgIndex = files.findIndex(({ name }) => name == "package.json")
+		const pkg = files[pkgIndex].content
 
-		options.extra.forEach(( ex ) => {
-			switch (ex) {
-				case "database": {
-					switch (options.database) {
-						case "mongodb": {
-							files[pkg].content.dependencies.mongodb = "^4.1.0"
-
-							files[srcIndex].files[srcLibIndex].files.push({
-								name: !options.typescript ? "mongodb.js" : "mongodb.ts",
-								content: !options.typescript ? misc.mongodb.js : misc.mongodb.ts
-							})
-
-							if (options.typescript) files[pkg].content.devDependencies["@types/mongodb"] = "^4.0.7"
-							break
-						}
-						case "mysql": {
-							files[pkg].content.dependencies.mysql = "^2.18.1"
-
-							files[srcIndex].files[srcLibIndex].files.push({
-								name: !options.typescript ? "mysql.js" : "mysql.ts",
-								content: !options.typescript ? misc.mysql.js : misc.mysql.ts
-							})
-
-							if (options.typescript) files[pkg].content.devDependencies["@types/mysql"] = "^2.15.19"
-							break
-						}
+		if (options.extra.includes("database")) {
+			switch (options.database) {
+				case "mongodb": {
+					pkg.dependencies = {
+						...pkg.dependencies,
+						mongodb: "^4.1.1"
 					}
+
+					files[srcIndex].children[srcLibIndex].children.push({
+						name: !options.typescript ? "mongodb.js" : "mongodb.ts",
+						content: !options.typescript ? misc.mongodb.js : misc.mongodb.ts
+					})
 					break
 				}
-				case "preprocess": {
-					if (!options.typescript) {
-						files[pkg].content.devDependencies["svelte-preprocess"] = "^4.7.4"
+				case "mysql": {
+					pkg.dependencies = {
+						...pkg.dependencies,
+						mysql: "^2.18.1"
 					}
+
+					files[srcIndex].children[srcLibIndex].children.push({
+						name: !options.typescript ? "mysql.js" : "mysql.ts",
+						content: !options.typescript ? misc.mysql.js : misc.mysql.ts
+					})
+
+					if (options.typescript) files[pkgIndex].content.devDependencies["@types/mysql"] = "^2.15.19"
 					break
 				}
 			}
-		})
+		}
+
+		if (options.extra.includes("preprocess")) {
+			if (!options.typescript) {
+				pkg.devDependencies["svelte-preprocess"] = "^4.7.4"
+			}
+		}
+
+		if (options.extra.includes("scripts") && options.scripts.length) {
+			const scripts = [{ name: "_utils.js", content: misc.scripts._utils }]
+
+			if (options.scripts.includes("build")) {
+				pkg.scripts = {
+					...pkg.scripts,
+					build: "npm run build:pre && npm run build:build && npm run build:post",
+					"build:pre": "node scripts/build.pre.js",
+					"build:build": "svelte-kit build --verbose",
+					"build:post": "node scripts/build.post.js"
+				}
+
+				pkg.devDependencies.colorette = "^1.4.0"
+
+				if (options["build.zip"]) {
+					pkg.devDependencies["7zip-min"] = "^1.3.3"
+				}
+
+				if (options["build.extractLicenses"]) {
+					pkg.devDependencies["@m4rch/extract-licenses"] = "^0.1.0"
+				}
+
+				scripts.push({
+					name: "build.pre.js",
+					content: misc.scripts.build.pre({ "7z": options["build.7z"], zip: options["build.zip"] })
+				})
+
+				scripts.push({
+					name: "build.post.js",
+					content: misc.scripts.build.post({
+						pkg,
+						"7z": options["build.7z"],
+						zip: options["build.zip"],
+						install: options["build.install"],
+						extractLicenses: options["build.extractLicenses"]
+					})
+				})
+			}
+
+			files.push({ name: "scripts", children: scripts })
+		}
+
+		files[pkgIndex].content = pkg
 	}
 
 	return files